+++ MovieAnalytics.API/DTOs/Analytics/MovieROIDto.cs
@@ -5,7 +5,7 @@
         public required string Title { get; set; }
         public decimal? Budget { get; set; }
         public decimal? GrossWorldWide { get; set; }
-        public decimal? ROI { get; set; }
+        public double? ROI { get; set; }
 
     }
 }
